% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_jobs.R
\name{gcva_automl_tabluar_training_job}
\alias{gcva_automl_tabluar_training_job}
\title{Constructs an AutoML Tabular Training Job}
\usage{
gcva_automl_tabluar_training_job(
  displayName,
  weightColumn = NULL,
  optimizationPredictionType = c("regression", "classification"),
  budgetMilliNodeHours = 1000,
  optimizationObjective = NULL,
  column_transformations = NULL
)
}
\arguments{
\item{displayName}{STRING Required. The user-defined name of this TrainingPipeline}

\item{weightColumn}{STRING column name of weight}

\item{optimizationPredictionType}{STRING Required. model type, one of: "regression", "classification"}

\item{budgetMilliNodeHours}{model budget, defaults to 1 hour or 1000}

\item{optimizationObjective}{objective to optimize for}

\item{column_transformations}{list of columns and column types, each pair a list}
}
\value{
"gcva_automlTabularTrainingJob" object
}
\description{
constructs a training job for running
}
