% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_jobs.R
\name{gcva_batch_prediction_job}
\alias{gcva_batch_prediction_job}
\title{Get a BatchPrediction object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.batchPredictionJobs/get}
\usage{
gcva_batch_prediction_job(locationId = gcva_region_get(), batchPredictionJob)
}
\arguments{
\item{locationId}{locationId of training pipeline}

\item{batchPredictionJob}{string Required. The name of the BatchPredictionJob resource.
Format: `projects/{project}/locations/{location}/batchPredictionJobs/{batchPredictionJob}`}
}
\value{
`BatchPredictionJob` object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.batchPredictionJobs
}
\description{
Get a BatchPrediction object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.batchPredictionJobs/get
}
