% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gcva_bucket_set}
\alias{gcva_bucket_set}
\title{Set staging bucket prefix}
\usage{
gcva_bucket_set(stagingBucket)
}
\arguments{
\item{stagingBucket}{The The Cloud Storage location where the training data is to be written to. In the given directory a new directory is created with name: dataset-<dataset-id>-<annotation-type>-<timestamp-of-training-call> where timestamp is in YYYY-MM-DDThh:mm:ss.sssZ ISO-8601 format. All training input data is written into that directory.
If the uri doesn't end with '/', a '/' will be automatically appended. The directory is created if it doesn't exist.}
}
\description{
Set staging bucket prefix
}
