% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_language_models.R
\name{gcva_ChatMessage}
\alias{gcva_ChatMessage}
\title{A chat message object}
\usage{
gcva_ChatMessage(content = content, auther = author)
}
\arguments{
\item{content:}{Content of the message.}

\item{author:}{Author of the message.}
}
\description{
A chat message object
}
