% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcva_create_tabluar_dataset}
\alias{gcva_create_tabluar_dataset}
\title{Creates a tabluar dataset}
\usage{
gcva_create_tabluar_dataset(
  projectId = gcva_project_get(),
  locationId = gcva_region_get(),
  displayName,
  gcsSource
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{displayName}{the name of the dataset that is shown in the interface.
The name can be up to 32 characters long and can consist only of ASCII
Latin letters A-Z and a-z, underscores (_), and ASCII digits 0-9.}
}
\description{
Creates a tabluar dataset
}
