% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_pipeline_jobs.R
\name{gcva_custom_container_training_job}
\alias{gcva_custom_container_training_job}
\title{Create custom container training job}
\usage{
gcva_custom_container_training_job(
  projectId = gcva_project_get(),
  locationId = gcva_region_get(),
  displayName,
  containerUri,
  command,
  serviceAccount = NULL,
  machineType = NULL,
  acceleratorType = NULL,
  acceleratorCount = NULL,
  replicaCount = NULL,
  args = NULL,
  modelDisplayName = NULL,
  predictSchemata = NULL,
  modelServingContainerImageUri = NULL,
  modelServingContainerCommand = NULL
)
}
\arguments{
\item{model_serving_container_command}{}
}
\description{
constructs a custom training job for running
}
