% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcva_dataset}
\alias{gcva_dataset}
\title{Get a dataset object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.datasets/get}
\usage{
gcva_dataset(locationId = gcva_region_get(), datasetName)
}
\arguments{
\item{locationId}{locationId of operation}

\item{datasetName}{the full name of the dataset object to get}
}
\value{
a Dataset object
}
\description{
Get a dataset object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.datasets/get
}
