% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gcva_deploy}
\alias{gcva_deploy}
\title{Deploy a model to an Endpoint}
\usage{
gcva_deploy(
  projectId = gcva_project_get(),
  locationId = gcva_region_get(),
  model = NULL,
  modelVersionId = NULL,
  endpoint = NULL,
  enableAccessLogging = NULL,
  trafficSplit = NULL,
  machineType = "n1-standard-4",
  minReplicaCount = 1,
  maxReplicaCount = NULL
)
}
\arguments{
\item{model}{Model object}

\item{endpoint}{Endpoint object}

\item{machineType}{STRING}
}
\value{
Endpoint object
}
\description{
Deploy a model to an Endpoint
}
\seealso{
Other Endpoints: 
\code{\link{gcva_create_endpoint}()},
\code{\link{gcva_delete_endpoint}()},
\code{\link{gcva_endpoint}()},
\code{\link{gcva_list_endpoints}()}
}
\concept{Endpoints}
