% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gcva_endpoint}
\alias{gcva_endpoint}
\title{Get an Endpoint}
\usage{
gcva_endpoint(locationId = gcva_region_get(), endpointName)
}
\arguments{
\item{endpointName}{}
}
\value{
a Endpoint object
}
\description{
Get an Endpoint
}
\seealso{
Other Endpoints: 
\code{\link{gcva_create_endpoint}()},
\code{\link{gcva_delete_endpoint}()},
\code{\link{gcva_deploy}()},
\code{\link{gcva_list_endpoints}()},
\code{\link{gcva_predict}()},
\code{\link{gcva_undeploy}()},
\code{\link{gcva_wait_for_model_deploy}()}
}
\concept{Endpoints}
