% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gcva_endpoint_create}
\alias{gcva_endpoint_create}
\title{Creates a Vertex AI Endpoint}
\usage{
gcva_endpoint_create(
  project = NULL,
  location = NULL,
  name = NULL,
  display_name = NULL,
  description = NULL,
  deployedModels = NULL,
  trafficSplit = NULL,
  etag = NULL,
  labels = NULL,
  createTime = NULL,
  updateTime = NULL,
  encryptionSpec = NULL,
  network = NULL,
  enablePrivateServiceConnect = NULL,
  modelDeploymentMonitoringJob = NULL,
  predictRequestResponseLoggingConfig = NULL
)
}
\arguments{
\item{displayName}{STRING Required. Required. The display name of the Endpoint. The name can be up to 128 characters long and can consist of any UTF-8 characters.}
}
\value{
operation object
https://cloud.google.com/vertex-ai/docs/reference/rest/Shared.Types/ListOperationsResponse#Operation
}
\description{
Creates an Endpoint Resource. Models are deployed into it, and afterwards Endpoint is called to obtain predictions and explanations.
}
