% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemini.R
\name{gcva_gemini_text}
\alias{gcva_gemini_text}
\title{Generate Text with Google Cloud Vertex AI Gemini Model}
\usage{
gcva_gemini_text(
  projectId = gcva_project_get(),
  locationId = gcva_region_get(),
  modelId = c("gemini-1.0-pro"),
  stream = FALSE,
  role = c("user"),
  prompt,
  data = NULL,
  mimeType = NULL,
  fileUri = NULL,
  startOffsetSeconds = NULL,
  startOffsetNanos = NULL,
  endOffsetSeconds = NULL,
  endOffsetNanos = NULL,
  functionDeclarations = NULL,
  harmCategory = c("HARM_CATEGORY_SEXUALLY_EXPLICIT", "HARM_CATEGORY_HATE_SPEECH",
    "HARM_CATEGORY_HARASSMENT", "HARM_CATEGORY_DANGEROUS_CONTENT"),
  harmThreshold = "BLOCK_ONLY_HIGH",
  candidateCount = NULL,
  temperature = 0.5,
  maxOutputTokens = NULL,
  topP = 1,
  topK = 10,
  stopSequences = NULL
)
}
\arguments{
\item{projectId}{The ID of the Google Cloud project. Default is retrieved using `gcva_project_get()`.}

\item{locationId}{The location ID for the AI model. Default is retrieved using `gcva_region_get()`.}

\item{modelId}{Character vector specifying the model ID, ALWAYS "gemini-1.0-pro".}

\item{stream}{Logical; whether to stream the output. ALWAYS FALSE}

\item{role}{Character vector specifying the user role, default is "user".}

\item{prompt}{Character string specifying the input prompt for text generation.}

\item{data}{Optional additional data for the request. UNUSED}

\item{mimeType}{The MIME type of the `data`. UNUSED}

\item{fileUri}{URI of the file containing data.}

\item{startOffsetSeconds}{Start offset in seconds for processing the input. UNUSED}

\item{startOffsetNanos}{Nanoseconds part of the start offset. UNUSED}

\item{endOffsetSeconds}{End offset in seconds for processing the input. UNUSED}

\item{endOffsetNanos}{Nanoseconds part of the end offset. UNUSED}

\item{functionDeclarations}{Additional function declarations for the request. UNUSED}

\item{harmCategory}{Categories of harmful content to be filtered, with defaults set.}

\item{harmThreshold}{Threshold for blocking harmful content, default is "BLOCK_ONLY_HIGH".}

\item{candidateCount}{Number of candidate responses to generate.}

\item{temperature}{Controls randomness in response generation, default is 0.5.}

\item{maxOutputTokens}{The maximum number of output tokens, with a maximum value of 8192.}

\item{topP}{Controls diversity of generated responses, default is 1.0.}

\item{topK}{Limits the number of high probability tokens considered for each step, default is 10.}

\item{stopSequences}{Sequences where the model should stop generating further tokens.}
}
\value{
A response object from the Google Cloud Vertex AI.
}
\description{
This function interfaces with Google Cloud Vertex AI to generate text using the Gemini model.
It allows for specifying various parameters like project ID, location ID, model ID, and more,
to customize the text generation process.
https://cloud.google.com/vertex-ai/docs/generative-ai/model-reference/gemini#sample_requests
}
\examples{
# Example usage:
response <- gcva_gemini_text(projectId = "your-project-id",
                             locationId = "your-location-id",
                             modelId = "gemini-1.0-pro",
                             prompt = "Hello, world!")

}
