% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gcva_get_op}
\alias{gcva_get_op}
\title{Poll an operationId
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.operations/get}
\usage{
gcva_get_op(locationId = gcva_region_get(), operationId)
}
\arguments{
\item{locationId}{locationId of operation}

\item{operationId}{operationId to poll}
}
\value{
an operation object
}
\description{
Poll an operationId
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.operations/get
}
