% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_language_models.R
\name{gcva_InputOutputTextPair}
\alias{gcva_InputOutputTextPair}
\title{TODO
A chat inputOutput text pair object}
\usage{
gcva_InputOutputTextPair(input_text, output_text)
}
\description{
TODO
A chat inputOutput text pair object
}
