% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcva_list_datasets}
\alias{gcva_list_datasets}
\title{Lists datasets in a project.}
\usage{
gcva_list_datasets(
  projectId = gcva_project_get(),
  locationId = gcva_region_get()
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}
}
\value{
a list of datasets
}
\description{
Lists datasets in a project.
}
