% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gcva_list_endpoints}
\alias{gcva_list_endpoints}
\title{Lists Vertex AI Endpoints within a specific GCP Project}
\usage{
gcva_list_endpoints(
  projectId = gcva_project_get(),
  locationId = gcva_region_get()
)
}
\description{
Lists Vertex AI Endpoints within a specific GCP Project
}
\seealso{
Other Endpoints: 
\code{\link{gcva_create_endpoint}()},
\code{\link{gcva_delete_endpoint}()},
\code{\link{gcva_endpoint}()}
}
\concept{Endpoints}
