% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcva_list_models}
\alias{gcva_list_models}
\title{Lists models in a project.}
\usage{
gcva_list_models(
  projectId = gcva_project_get(),
  locationId = gcva_region_get()
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}
}
\description{
Lists models in a project.
}
