% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gcva_predict}
\alias{gcva_predict}
\title{Predict}
\usage{
gcva_predict(locationId = gcva_region_get(), endpoint, instances)
}
\arguments{
\item{locationId}{locationId of operation}

\item{endpoint}{an `Endpoint` resource name "projects/{project}/locations/{location}/endpoints/{endpoint}"}

\item{instances}{a json object for online prediction request}
}
\description{
Predict
}
\seealso{
Other Endpoints: 
\code{\link{gcva_create_endpoint}()},
\code{\link{gcva_delete_endpoint}()},
\code{\link{gcva_deploy}()},
\code{\link{gcva_endpoint}()},
\code{\link{gcva_list_endpoints}()},
\code{\link{gcva_undeploy}()},
\code{\link{gcva_wait_for_model_deploy}()}
}
\concept{Endpoints}
