% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gcva_region_set}
\alias{gcva_region_set}
\title{Set the region}
\usage{
gcva_region_set(
  region = c("us-central1", "us-east1", "us-east4", "us-west1",
    "northamerica-northeast1", "europe-west1", "europe-west2", "europe-west4",
    "asia-east1", "asia-northeast1", "asia-northeast3", "asia-southeast1",
    "australia-southeast1")
)
}
\arguments{
\item{region}{Region for the endpoint, full list of service endpoints
[here](https://cloud.google.com/vertex-ai/docs/reference/rest#service-endpoint)}
}
\description{
Can also use environment argument GCVA_DEFAULT_REGION
}
\details{
Current
}
