% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_pipeline_jobs.R
\name{gcva_trainingPipeline}
\alias{gcva_trainingPipeline}
\title{Get a Training Pipeline object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.trainingPipelines/get}
\usage{
gcva_trainingPipeline(locationId = gcva_region_get(), trainingPipelineName)
}
\arguments{
\item{locationId}{locationId of training pipeline}

\item{trainingPipelineName}{trainingPipeline full name or uri}
}
\value{
a training Pipeline object
}
\description{
Get a Training Pipeline object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.trainingPipelines/get
}
