% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gcva_undeploy}
\alias{gcva_undeploy}
\title{Undeploy Model from Endpoint}
\usage{
gcva_undeploy(
  projectId = gcva_project_get(),
  locationId = gcva_region_get(),
  endpoint
)
}
\arguments{
\item{endpoint}{string Required. Endpoint object `gcva_endpoint` or
The name of the Endpoint resource from which to undeploy a Model.
Format: projects/{project}/locations/{location}/endpoints/{endpoint}}
}
\description{
Undeploy Model from Endpoint
}
