% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_prediction_jobs.R
\name{gcva_wait_for_batch_prediction_job}
\alias{gcva_wait_for_batch_prediction_job}
\title{Wait for a BatchPredctionJob operation
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/JobState}
\usage{
gcva_wait_for_batch_prediction_job(
  locationId = gcva_region_get(),
  batchPredictionJob,
  wait = 300
)
}
\arguments{
\item{locationId}{locationId of training pipeline}

\item{wait}{INTEGER number of seconds to wait between checks. Default is 5minutes}

\item{batchPredictionJobName}{Resource name of the BatchPredictionJob.}
}
\value{
batchPredictionJob object
}
\description{
Wait for a BatchPredctionJob operation
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/JobState
}
