% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gcva_wait_for_model_deploy}
\alias{gcva_wait_for_model_deploy}
\title{Wait for an Model to deploy}
\usage{
gcva_wait_for_model_deploy(
  locationId = gcva_region_get(),
  operation,
  wait = 10
)
}
\arguments{
\item{locationId}{locationId of operation}

\item{operation}{A operation object}

\item{wait}{the number of seconds to wait between checks
Use this function to do a loop to check progress of a operation running}
}
\value{
After a while, a completed model deployment operation
}
\description{
Wait for an Model to deploy
}
