% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gcva_wait_for_op}
\alias{gcva_wait_for_op}
\title{Wait for an Operation}
\usage{
gcva_wait_for_op(locationId = gcva_region_get(), operation, wait = 2)
}
\arguments{
\item{locationId}{locationId of operation}

\item{operation}{A operation object}

\item{wait}{the number of seconds to wait between checks
Use this function to do a loop to check progress of a operation running}
}
\value{
After a while, a completed operation
}
\description{
Wait for an Operation
}
