% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCloudVertexAIR.R
\docType{package}
\name{googleCloudVertexAIR}
\alias{googleCloudVertexAIR-package}
\alias{googleCloudVertexAIR}
\title{Interact with Google Cloud Vertex AI with R}
\description{
A tool to interact with the Google Cloud Vertex AI API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/justinjm/googleCloudVertexAIR}
  \item Report bugs at \url{https://github.com/justinjm/googleCloudVertexAIR/issues}
}

}
\author{
\strong{Maintainer}: Justin Marciszewski \email{justin@harborislandanalytics.com} (\href{https://orcid.org/0000-0002-8251-4982}{ORCID})

}
