% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_jobs.R
\name{gcva_automl_tabluar_training_job}
\alias{gcva_automl_tabluar_training_job}
\title{Constructs an AutoML Tabular Training Job}
\usage{
gcva_automl_tabluar_training_job(
  displayName,
  weightColumn = NULL,
  optimizationPredictionType = c("regression", "classification"),
  budgetMilliNodeHours = 1000,
  optimizationObjective = NULL,
  column_transformations = NULL
)
}
\arguments{
\item{displayName}{STRING Required. The user-defined name of this TrainingPipeline.}

\item{column_transformations}{}
}
\description{
Constructs an AutoML Tabular Training Job
}
