% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gcva_delete_dataset}
\alias{gcva_delete_dataset}
\title{Deletes a dataset}
\usage{
gcva_delete_dataset(
  projectId = gcva_project_get(),
  locationId = gcva_region_get(),
  displayName = NULL,
  dataset
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{displayName}{the name of the dataset that is shown in the interface.}

\item{dataset}{dataset object returned by gcva_dataset function}
}
\description{
Deletes a dataset
}
