% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gcva_model}
\alias{gcva_model}
\title{Get a Model object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.models/get}
\usage{
gcva_model(locationId = gcva_region_get(), model)
}
\arguments{
\item{locationId}{locationId of operation}

\item{model}{the full name of the model object to get OR
a trainingPipelineJob object}
}
\value{
a Model object
}
\description{
Get a Model object
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.models/get
}
\seealso{
Other Models: 
\code{\link{gcva_list_models}()}
}
\concept{Models}
