% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_pipeline_jobs.R
\name{gcva_run_job}
\alias{gcva_run_job}
\title{Executes an training job
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.trainingPipelines/create}
\usage{
gcva_run_job(
  projectId = gcva_project_get(),
  locationId = gcva_region_get(),
  job,
  dataset,
  targetColumn = NULL,
  trainingFractionSplit = 0.8,
  validationFractionSplit = 0.1,
  testFractionSplit = 0.1,
  modelDisplayName,
  disableEarlyStopping = FALSE,
  machineType = NULL,
  sync = TRUE
)
}
\arguments{
\item{projectId}{GCP project id}

\item{locationId}{location of GCP resources}

\item{job}{a job object of:  "gcva_automlTabularTrainingJob" OR TBD}

\item{dataset}{a vertex ai dataset object with class "gcva_dataset"}

\item{targetColumn}{STRING full column name of the model target}

\item{trainingFractionSplit}{decimal, percentage of dataset to use for training model during training}

\item{validationFractionSplit}{decimal, percentage of dataset to use for validating model during training}

\item{testFractionSplit}{decimal, percentage of dataset to use for testing model during training}

\item{modelDisplayName}{display name of the model}

\item{disableEarlyStopping}{disable early stopping default FALSE}

\item{sync}{If set to TRUE, the call will block while waiting for the asynchronous batch job to complete.}
}
\value{
\href{https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.trainingPipelines#TrainingPipeline}{trainingPipeline} object
}
\description{
Executes an training job
https://cloud.google.com/vertex-ai/docs/reference/rest/v1/projects.locations.trainingPipelines/create
}
